+++ Course Work/Course Work/DetermineCalc.h
@@ -1,5 +1,5 @@
 #pragma once
 
-int det(int**, int); // Прототип функції обчислення визначника
-int AlgebralAdditionCalc(int**, const int&, const int&); // Прототип функції обчислення алгебраїного доповнення заданого елементу
-int** matrixCut(int**, const int&, const int&); // Прототип функції уворення матриці (N-1)-го порядку
+double det(double**, int); // Прототип функції обчислення визначника
+double AlgebralAdditionCalc(double**, const int&, const int&); // Прототип функції обчислення алгебраїного доповнення заданого елементу
+double** matrixCut(double**, const int&, const int&); // Прототип функції уворення матриці (N-1)-го порядку
